chore(desktop): optimize test suite and extract AudioNormalizationSection
Switch Vitest from jsdom to happy-dom with isolate:false for faster module
reuse. Tag 7 pure-logic test files with @vitest-environment node to skip
DOM overhead. Replace userEvent with fireEvent in simple click tests.

Extract AudioNormalizationSection (341 lines) from SettingsPage, reducing
it from 741 to 435 lines. Move 7 normalization visibility tests to
isolated component test file (9 tests, 266ms vs 1687ms full page render).

Test suite: 283 tests in ~5s (was 42s with jsdom + isolation).
