+++ crates/rdlp-desktop/package.json
@@ -43,6 +43,7 @@
     "@vitejs/plugin-react": "^4.3.0",
     "@vitest/coverage-v8": "^4.0.18",
     "cypress": "^15.10.0",
+    "happy-dom": "^20.8.3",
     "jsdom": "^28.1.0",
     "typescript": "^5.4.5",
     "vite": "^5.4.0",
