+++ crates/rdlp-desktop/src/components/utils/tableHelpers.test.ts
@@ -1,3 +1,4 @@
+// @vitest-environment node
 // Unit tests for tableHelpers.ts — groupRows and optionsSummary.
 
 import { describe, test, expect } from "vitest";
