+++ crates/rdlp-desktop/src/query/queryKeys.test.ts
@@ -1,3 +1,4 @@
+// @vitest-environment node
 // Unit tests for queryKeys.ts — deterministic key generation.
 
 import { describe, test, expect } from "vitest";
