+++ crates/rdlp-desktop/vitest.config.ts
@@ -11,10 +11,11 @@ export default defineConfig({
     },
     test: {
         globals: true,
-        environment: "jsdom",
+        environment: "happy-dom",
         setupFiles: ["./src/test/setup.ts"],
         include: ["src/**/*.{test,spec}.{ts,tsx}"],
         pool: "threads",
+        isolate: false,
         coverage: {
             provider: "v8",
             reporter: ["text", "json", "html"],
